Refactor IRoomList
The INotifyCollectionChanged is not suitable to cases where the collection is
changed on a different thread (there is no way to get a reliable "starting point").
In general the event model is difficult to combine with multithreading, so
simple polling is probably better.

Also exposed the list refresh, so callers that are not bound to a predefined
update loop can synchronize to the internal updates instead.