Use same family as the broadcast address
The default family used by the Socket constructor is system-dependent and might
be different if the same app is compiled on different platforms (e.g. IPv6 on
Windows, IPv4 on Android). So we force the address family to the one of the
broadcast address, that will be the same across platforms (provided that the
broadcast address is the same).